Move setting up sceneSettings to scene-controller, moving a lot of stuff around
