+++ src/fontra/client/core/glyph-lines.js
@@ -1,0 +1,99 @@
+import {
+  getSuggestedGlyphName,
+  getUnicodeFromGlyphName,
+  parseClipboard,
+} from "./server-utils.js";
+
+export async function glyphLinesFromText(text, characterMap, glyphMap) {
+  const glyphLines = [];
+  for (const line of text.split(/\r?\n/)) {
+    glyphLines.push(await glyphNamesFromText(line, characterMap, glyphMap));
+  }
+  return glyphLines;
+}
+
+const glyphNameRE = /[//\s]/g;
+
+async function glyphNamesFromText(text, characterMap, glyphMap) {
+  const glyphNames = [];
+  for (let i = 0; i < text.length; i++) {
+    let glyphName;
+    let char = text[i];
+    if (char == "/") {
+      i++;
+      if (text[i] == "/") {
+        glyphName = characterMap[char.charCodeAt(0)];
+      } else {
+        glyphNameRE.lastIndex = i;
+        glyphNameRE.test(text);
+        let j = glyphNameRE.lastIndex;
+        if (j == 0) {
+          glyphName = text.slice(i);
+          i = text.length - 1;
+        } else {
+          j--;
+          glyphName = text.slice(i, j);
+          if (text[j] == "/") {
+            i = j - 1;
+          } else {
+            i = j;
+          }
+        }
+        char = undefined;
+        for (const codePoint of glyphMap[glyphName] || []) {
+          if (characterMap[codePoint] === glyphName) {
+            char = String.fromCodePoint(codePoint);
+            break;
+          }
+        }
+        if (!char && !glyphMap[glyphName]) {
+          // Glyph doesn't exist in the font, try to find a unicode value
+          const codePoint = await getUnicodeFromGlyphName(glyphName);
+          if (codePoint) {
+            char = String.fromCodePoint(codePoint);
+          }
+        }
+      }
+    } else {
+      const charCode = text.codePointAt(i);
+      glyphName = characterMap[charCode];
+      if (charCode >= 0x10000) {
+        i++;
+      }
+      char = String.fromCodePoint(charCode);
+    }
+    if (glyphName !== "") {
+      let isUndefined = false;
+      if (!glyphName && char) {
+        glyphName = await getSuggestedGlyphName(char.codePointAt(0));
+        isUndefined = true;
+      }
+      glyphNames.push({
+        character: char,
+        glyphName: glyphName,
+        isUndefined: isUndefined,
+      });
+    }
+  }
+  return glyphNames;
+}
+
+export function textFromGlyphLines(glyphLines) {
+  const textLines = [];
+  for (const glyphLine of glyphLines) {
+    let textLine = "";
+    for (let i = 0; i < glyphLine.length; i++) {
+      const glyphInfo = glyphLine[i];
+      if (glyphInfo.character) {
+        textLine += glyphInfo.character;
+      } else {
+        textLine += "/" + glyphInfo.glyphName;
+        if (glyphLine[i + 1]?.character) {
+          textLine += " ";
+        }
+      }
+    }
+    textLines.push(textLine);
+  }
+  return textLines.join("\n");
+}
